requirements.txt
# Core dependencies
boto3>=1.28.0
influxdb-client>=1.36.0
python-dotenv>=1.0.0

# Optional: For Kafka integration
# kafka-python>=2.0.2

# Development dependencies (optional)
# jupyter>=1.0.0
# pandas>=2.0.0
# matplotlib>=3.7.0
